Rename depth_measures to depth
