+++ docs/modules/exploratory.rst
@@ -9,4 +9,4 @@ and visualize functional data.
    :caption: Modules:
    
    exploratory/visualization
-   exploratory/depth_measures
+   exploratory/depth
