+++ skfda/exploratory/__init__.py
@@ -1,3 +1,3 @@
 from . import visualization
 from . import stats
-from . import depth_measures
+from . import depth
