+++ skfda/exploratory/depth.py
